Fix below error when exiting by setting state null.
(bt01_helloworld.py:31151): GStreamer-CRITICAL **:
Trying to dispose element playbin20, but it is in PLAYING instead of the NULL state.
You need to explicitly set elements to the NULL state before
dropping the final reference, to allow them to clean up.
This problem may also be caused by a refcounting bug in the
application or some element.